+++ packages/cli.cmd.typescript/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osdk/cli.cmd.typescript
 
+## 0.84.0
+
+### Patch Changes
+
+- @osdk/generator@2.59.0
+- @osdk/cli.common@0.84.0
+
+## 0.83.0
+
+### Patch Changes
+
+- Updated dependencies [ab557b4]
+- Updated dependencies [90cce42]
+  - @osdk/generator@2.58.0
+  - @osdk/cli.common@0.83.0
+
 ## 0.82.0
 
 ### Patch Changes