+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/cli
 
+## 0.84.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.57.0
+
+## 0.83.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.56.0
+
 ## 0.82.0
 
 ### Patch Changes