+++ packages/create-app.template-packager/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/create-app.template-packager
 
+## 2.59.0
+
+### Minor Changes
+
+- b913397: Remove the create-widget minimal-react template which is now replaced by the react template rendered without OSDK. The minimal-react template is still available via CLI selection for compatibility and is simply an alias of the react template rendered without OSDK.
+- 7c4c517: Use minimal react template structure for react template
+
+## 2.58.0
+
+### Minor Changes
+
+- 8718d53: create-widget.template.react.v2 can be used with or without an osdk package
+
 ## 2.57.0
 
 ## 2.56.0