+++ packages/example-generator/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @osdk/example-generator
 
+## 0.48.0
+
+### Minor Changes
+
+- b913397: Remove the create-widget minimal-react template which is now replaced by the react template rendered without OSDK. The minimal-react template is still available via CLI selection for compatibility and is simply an alias of the react template rendered without OSDK.
+- 7c4c517: Use minimal react template structure for react template
+
+### Patch Changes
+
+- Updated dependencies [b913397]
+- Updated dependencies [7c4c517]
+  - @osdk/create-widget@3.57.0
+  - @osdk/create-app@2.59.0
+
+## 0.47.0
+
+### Minor Changes
+
+- 8718d53: create-widget.template.react.v2 can be used with or without an osdk package
+
+### Patch Changes
+
+- Updated dependencies [8718d53]
+  - @osdk/create-widget@3.56.0
+  - @osdk/create-app@2.58.0
+
 ## 0.46.0
 
 ### Patch Changes