+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @osdk/widget.api
 
+## 3.57.0
+
+### Patch Changes
+
+- Updated dependencies [4ae6d8b]
+- Updated dependencies [1aff7f4]
+  - @osdk/api@2.59.0
+
+## 3.56.0
+
+### Patch Changes
+
+- @osdk/api@2.58.0
+
 ## 3.55.0
 
 ### Patch Changes